reorganizes file structure to simplify files
